refactor: extract store source metadata helpers
