+++ src/codex_usage_tracker/store.py
@@ -3,12 +3,9 @@
 from __future__ import annotations
 
 import csv
-import hashlib
-import json
 import sqlite3
 from collections.abc import Iterable, Iterator
 from contextlib import contextmanager, suppress
-from dataclasses import dataclass
 from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any
@@ -16,13 +13,10 @@ from typing import Any
 from codex_usage_tracker.models import RefreshResult, UsageEvent
 from codex_usage_tracker.parser import (
     PARSER_DIAGNOSTIC_KEYS,
-    ParserState,
     compact_parser_diagnostics,
     find_session_logs,
     load_session_index,
     parse_usage_events_from_file_with_state,
-    parser_state_from_json,
-    parser_state_to_json,
 )
 from codex_usage_tracker.paths import DEFAULT_CODEX_HOME, DEFAULT_DB_PATH
 from codex_usage_tracker.projects import apply_project_privacy_to_rows, validate_privacy_mode
@@ -46,18 +40,14 @@ from codex_usage_tracker.store_schema import (
     SchemaMigrationError,
     init_db,
 )
+from codex_usage_tracker.store_sources import (
+    ParsedSourceFile,
+    source_logs_requiring_parse,
+    upsert_source_file_metadata,
+)
 
 EVENT_COLUMNS = list(USAGE_EVENT_COLUMN_NAMES)
 __all__ = ["EVENT_COLUMNS", "SCHEMA_VERSION", "SchemaMigrationError", "init_db"]
-
-
-@dataclass(frozen=True)
-class SourceParsePlan:
-    path: Path
-    start_byte: int = 0
-    start_line: int = 0
-    initial_state: ParserState | None = None
-    replace_existing: bool = True
 
 
 def refresh_usage_index(
@@ -69,10 +59,12 @@ def refresh_usage_index(
 
     logs = find_session_logs(codex_home=codex_home, include_archived=include_archived)
     session_index = load_session_index(codex_home)
-    parse_plans = _source_logs_requiring_parse(logs, db_path=db_path)
+    with connect(db_path) as conn:
+        init_db(conn)
+        parse_plans = source_logs_requiring_parse(conn, logs)
     stats: dict[str, int] = {}
     events: list[UsageEvent] = []
-    parsed_files: list[tuple[Path, list[UsageEvent], dict[str, int], ParserState]] = []
+    parsed_files: list[ParsedSourceFile] = []
     for plan in parse_plans:
         file_stats: dict[str, int] = {}
         parsed_file = parse_usage_events_from_file_with_state(
@@ -256,181 +248,19 @@ def schema_state(db_path: Path = DEFAULT_DB_PATH) -> dict[str, Any]:
     }
 
 
-def _source_logs_requiring_parse(
-    logs: Iterable[Path],
-    *,
-    db_path: Path,
-) -> list[SourceParsePlan]:
-    paths = list(logs)
-    if not paths:
-        return []
-    changed: list[SourceParsePlan] = []
-    with connect(db_path) as conn:
-        init_db(conn)
-        for path in paths:
-            metadata = _source_file_metadata(path)
-            if metadata is None:
-                continue
-            row = conn.execute(
-                """
-                SELECT size_bytes, mtime_ns, parsed_until_line
-                    , parsed_until_byte, parser_state_json
-                FROM source_files
-                WHERE source_file = ?
-                """,
-                (str(path),),
-            ).fetchone()
-            if row is None:
-                changed.append(SourceParsePlan(path=path))
-                continue
-            previous_size = int(row["size_bytes"])
-            previous_mtime_ns = int(row["mtime_ns"])
-            previous_byte = int(row["parsed_until_byte"])
-            previous_line = int(row["parsed_until_line"])
-            previous_state = parser_state_from_json(row["parser_state_json"])
-            if previous_state is None:
-                changed.append(SourceParsePlan(path=path))
-                continue
-            if (
-                previous_size == metadata["size_bytes"]
-                and previous_mtime_ns == metadata["mtime_ns"]
-            ):
-                continue
-            if metadata["size_bytes"] > previous_size and 0 < previous_byte <= previous_size:
-                changed.append(
-                    SourceParsePlan(
-                        path=path,
-                        start_byte=previous_byte,
-                        start_line=previous_line,
-                        initial_state=previous_state,
-                        replace_existing=False,
-                    )
-                )
-                continue
-            changed.append(SourceParsePlan(path=path))
-    return changed
-
-
 def record_source_file_metadata(
     db_path: Path = DEFAULT_DB_PATH,
     *,
-    parsed_files: Iterable[tuple[Path, list[UsageEvent], dict[str, int], ParserState]],
+    parsed_files: Iterable[ParsedSourceFile],
 ) -> None:
     """Record metadata for source files parsed during refresh."""
 
     parsed = list(parsed_files)
     if not parsed:
         return
-    indexed_at = datetime.now(timezone.utc).replace(microsecond=0).isoformat()
-    rows: list[dict[str, Any]] = []
-    for path, events, diagnostics, parser_state in parsed:
-        metadata = _source_file_metadata(path)
-        if metadata is None:
-            continue
-        latest_event = max(
-            events,
-            key=lambda event: (
-                event.event_timestamp,
-                event.cumulative_total_tokens,
-                event.line_number,
-                event.record_id,
-            ),
-            default=None,
-        )
-        rows.append(
-            {
-                "source_file_id": _source_file_id(path),
-                "source_file": str(path),
-                "source_file_hash": _source_file_hash(path),
-                "is_archived": int(metadata["is_archived"]),
-                "size_bytes": int(metadata["size_bytes"]),
-                "mtime_ns": int(metadata["mtime_ns"]),
-                "parsed_until_line": _count_lines(path),
-                "parsed_until_byte": int(metadata["size_bytes"]),
-                "latest_record_id": (
-                    latest_event.record_id
-                    if latest_event
-                    else parser_state.latest_record_id
-                ),
-                "latest_event_timestamp": (
-                    latest_event.event_timestamp
-                    if latest_event
-                    else parser_state.latest_event_timestamp
-                ),
-                "parser_adapter": "codex-jsonl-v1",
-                "parser_diagnostics_json": json.dumps(
-                    compact_parser_diagnostics(diagnostics),
-                    sort_keys=True,
-                ),
-                "parser_state_json": parser_state_to_json(parser_state),
-                "last_indexed_at": indexed_at,
-            }
-        )
-    if not rows:
-        return
-    columns = [
-        "source_file_id",
-        "source_file",
-        "source_file_hash",
-        "is_archived",
-        "size_bytes",
-        "mtime_ns",
-        "parsed_until_line",
-        "parsed_until_byte",
-        "latest_record_id",
-        "latest_event_timestamp",
-        "parser_adapter",
-        "parser_diagnostics_json",
-        "parser_state_json",
-        "last_indexed_at",
-    ]
-    placeholders = ", ".join("?" for _column in columns)
-    update_clause = ", ".join(
-        f"{column}=excluded.{column}" for column in columns if column != "source_file_id"
-    )
     with connect(db_path) as conn:
         init_db(conn)
-        conn.executemany(
-            (
-                f"INSERT INTO source_files ({', '.join(columns)}) "
-                f"VALUES ({placeholders}) "
-                f"ON CONFLICT(source_file_id) DO UPDATE SET {update_clause}"
-            ),
-            [[row[column] for column in columns] for row in rows],
-        )
-
-
-def _source_file_metadata(path: Path) -> dict[str, int] | None:
-    try:
-        stat = path.stat()
-    except OSError:
-        return None
-    return {
-        "size_bytes": int(stat.st_size),
-        "mtime_ns": int(stat.st_mtime_ns),
-        "is_archived": _is_archived_source_file(path),
-    }
-
-
-def _is_archived_source_file(path: Path) -> int:
-    normalized = str(path).replace("\\", "/")
-    return int("/archived_sessions/" in normalized or normalized.startswith("archived_sessions/"))
-
-
-def _source_file_id(path: Path) -> str:
-    return _source_file_hash(path)
-
-
-def _source_file_hash(path: Path) -> str:
-    return hashlib.sha256(str(path).encode("utf-8")).hexdigest()
-
-
-def _count_lines(path: Path) -> int:
-    try:
-        with path.open("rb") as handle:
-            return sum(1 for _line in handle)
-    except OSError:
-        return 0
+        upsert_source_file_metadata(conn, parsed_files=parsed)
 
 
 def upsert_usage_events(
